Add .gitignore to saved_games folder to push it
Pushing empty directories isn't possible, but the saved_games directory
is required for the save/load functionality to work correctly. Adding
a .gitignore file to this folder serves the dual purpose of allowing
the folder to be committed while also preventing any test saves that
wind up in the folder from being committed.